Improve memory management
Without the patch it is only possible to specify how much memory each
thread that is using the SprayList will be able to allocate. This is
problematic for the single source shortest paths benchmark (bin/sssp)
as it makes it impossible to run the benchmark with a single thread
and many threads using the same binary. The benchmark would either run
out of memory when a single thread is used or there would not be
enough memory on the machine when many threads are used.

With the patch one can specify how much memory all threads that are
using the SprayList can allocated together (the #define
SSALLOC_SIZE_ALL in include/ssalloc.h). The patch makes sure that this
amount of memory is divided between the threads.